requirements.txt
albumentations==1.1.0
apex==0.9.10dev
fvcore==0.1.5.post20221221
h5py==3.13.0
imgaug==0.4.0
matplotlib==3.10.3
medpy==0.5.2
monai==1.4.0
numpy==2.3.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.1
Pillow==11.3.0
ptflops==0.7.4
scipy==1.16.0
SimpleITK==2.4.1
SimpleITK==2.5.2
tensorboardX==2.6.2.2
tensorboardX==2.6.4
thop==0.1.1.post2209072238
timm==1.0.16
torch==2.7.0
torchvision==0.22.0
tqdm==4.65.2
triton==3.3.0
